Moving examples back to root examples folder;
